stock_usability: add button to access stock moves before button to access product moves, because stock moves are more important than product moves
