--- stock_usability/models/product_template.py
@@ -1,15 +1,0 @@
-# Copyright 2016-2020 Akretion France
-# @author: Alexis de Lattre <alexis.delattre@akretion.com>
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
-from odoo import fields, models
-
-
-class ProductTemplate(models.Model):
-    _inherit = 'product.template'
-
-    tracking = fields.Selection(tracking=True)
-    sale_delay = fields.Float(tracking=True)
-    # the 'stock' module adds 'product' in type...
-    # but forgets to make it the default
-    type = fields.Selection(default='product')
